requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
importlib-metadata==5.1.0
Markdown==3.4.1
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
uuid==1.30
zipp==3.11.0
